update connect withdraw function
