+++ package.json
@@ -1,7 +1,6 @@
 {
-  "name": "wakkanay-browser-wallet",
+  "name": "gazelle-wallet",
   "version": "0.1.0",
-  "private": true,
   "scripts": {
     "dev": "NODE_OPTIONS=--max_old_space_size=4096 next dev src -p 5001",
     "build": "next build src",
